+++ projects/tests/dbs/mongo.spec.ts
@@ -46,7 +46,7 @@ describe.skipIf(!mongoConnectionString)('mongo with Transaction', () => {
   )
   it('work with native mongo and condition', async () => {
     const repo = await entityWithValidations.create4RowsInDp(createEntity)
-    const mongo = MongoDataProvider.getDb(remult)
+    const mongo = MongoDataProvider.getDb(remult.dataProvider)
     const r = await (
       await mongo.db.collection(repo.metadata.dbName!)
     ).countDocuments(
@@ -56,7 +56,7 @@ describe.skipIf(!mongoConnectionString)('mongo with Transaction', () => {
   })
   it('work with native mongo', async () => {
     const repo = await entityWithValidations.create4RowsInDp(createEntity)
-    const mongo = MongoDataProvider.getDb(remult)
+    const mongo = MongoDataProvider.getDb(remult.dataProvider)
     const r = await (
       await mongo.db.collection(repo.metadata.dbName!)
     ).countDocuments()