+++ projects/tests/dbs/shared-tests/test-knex.ts
@@ -67,7 +67,7 @@ export function knexTests(
 
   it('knex with filter', async () => {
     const repo = await entityWithValidations.create4RowsInDp(createEntity)
-    const knex = KnexDataProvider.getDb(remult)
+    const knex = KnexDataProvider.getDb(remult.dataProvider)
     const e = await dbNamesOf(repo)
     expect(
       getKnexCount(
@@ -79,7 +79,7 @@ export function knexTests(
   })
   it('work with native knex3', async () => {
     const repo = await entityWithValidations.create4RowsInDp(createEntity)
-    const knex = KnexDataProvider.getDb(remult)
+    const knex = KnexDataProvider.getDb(remult.dataProvider)
     const t = await dbNamesOf(repo)
     const r = await knex((await t).$entityName).select(t.myId, t.name)
     expect(r.length).toBe(4)
@@ -87,14 +87,14 @@ export function knexTests(
 
   it('work with native knex', async () => {
     const repo = await entityWithValidations.create4RowsInDp(createEntity)
-    const knex = KnexDataProvider.getDb(remult)
+    const knex = KnexDataProvider.getDb(remult.dataProvider)
     const r = await knex(repo.metadata.dbName!).count()
     expect(getKnexCount(r)).toBe(4)
   })
   it('work with native knex2', async () => {
     const repo = await entityWithValidations.create4RowsInDp(createEntity)
     await remult.dataProvider.transaction(async (db) => {
-      const sql = KnexDataProvider.getDb(new Remult(db))
+      const sql = KnexDataProvider.getDb(db)
       const r = await sql(repo.metadata.dbName!).count()
       expect(getKnexCount(r)).toBe(4)
     })
@@ -138,7 +138,7 @@ export function knexTests(
     expect((await s.find()).map((x) => x.myId)).toEqual([1, 2, 3, 4])
   })
   it('test sql expression', async () => {
-    let knex = KnexDataProvider.getDb(remult)
+    let knex = KnexDataProvider.getDb(remult.dataProvider)
 
     @Entity('testSqlExpression')
     class testSqlExpression {